Pass grid size to meta balls rather than no. of grid points.

+++ coot/moorhen-wrappers.cc
@@ -136,7 +136,7 @@ struct moorhen_hbond {
 
 };
 
-coot::simple_mesh_t GenerateMoorhenMetaBalls(mmdb::Manager *molHnd, const std::string &cid_str, int gridSize, float radius, float isoLevel);
+coot::simple_mesh_t GenerateMoorhenMetaBalls(mmdb::Manager *molHnd, const std::string &cid_str, float gridSize, float radius, float isoLevel);
 coot::instanced_mesh_t DrawSugarBlocks(mmdb::Manager *molHnd, const std::string &cid_str);
 bool isSugar(const std::string &resName);
 
@@ -151,7 +151,7 @@ class molecules_container_js : public molecules_container_t {
             return DrawSugarBlocks(mol,cid_str);
         }
 
-        coot::simple_mesh_t DrawMoorhenMetaBalls(int imol, const std::string &cid_str, int gridSize, float radius, float isoLevel) {
+        coot::simple_mesh_t DrawMoorhenMetaBalls(int imol, const std::string &cid_str, float gridSize, float radius, float isoLevel) {
             mmdb::Manager *mol = get_mol(imol);
             return GenerateMoorhenMetaBalls(mol,cid_str,gridSize,radius,isoLevel);
         }
